requirements.txt
fastapi
pydantic
langchain
langchain-community
langchain-text-splitters
langchain-core
ollama
python-dotenv
langchain-openai
langchain-ollama
langchain-groq
langchain-chroma 